Create separate drawer component
