+++ frontend/src/components/createWorkflow/CreateWorkflow.jsx
@@ -1,41 +1,21 @@
 import React, { useContext, useState } from "react";
-import { Paper } from "@mui/material";
-import { StepperDiv } from "utils/Style/WorkflowStyle";
 
 // Create workflow
 import { VoteInfoContext } from "context/VoteInfoContext";
-import Stepper from "components/base/Stepper";
 import SelectParticipants from "components/createWorkflow/SelectParticipants";
 import SelectCandidates from "components/createWorkflow/SelectCandidates";
 import DatePicker from "components/createWorkflow/DatePicker";
 import ReviewDetails from "components/createWorkflow/ReviewDetails";
 import Payment from "components/createWorkflow/Payment";
-import Toolbar from "components/base/Toolbar";
-
-import { AppBar, Box, Drawer, Divider } from "@mui/material";
-import { styled } from "@mui/system";
 
 import { MINUTES_DELAY, DELAY } from "constants";
 import isSameDate from "utils/createWorkflow/IsSameDate";
 
-const drawerWidth = 240;
-
-const StyledDrawer = styled(Drawer)(({ theme }) => ({
-  "& .MuiDrawer-paper": {
-    boxSizing: "border-box",
-    width: drawerWidth,
-    backgroundColor: theme.palette.background.default,
-  },
-}));
+import WorkflowDrawer from "components/base/Drawer";
 
 const CreateVoteWorkflow = () => {
-  const [mobileOpen, setMobileOpen] = useState(false);
   const [voteInfo, setVoteInfo] = useContext(VoteInfoContext);
   const [dateValue, setDateValue] = useState(null);
-
-  const handleDrawerToggle = () => {
-    setMobileOpen(!mobileOpen);
-  };
 
   const steps = [
     "Select Voting Participants",
@@ -45,20 +25,6 @@ const CreateVoteWorkflow = () => {
     "Review Details",
     "Payment & Title",
   ];
-
-  const drawer = (
-    <div>
-      <AppBar style={{ height: "64px" }}>
-        <Toolbar />
-      </AppBar>
-      <Divider />
-      <StepperDiv>
-        <Stepper steps={steps} />
-      </StepperDiv>
-    </div>
-  );
-
-  const container = window.document.body;
 
   const handleNext = () => {
     // Start date
@@ -108,110 +74,55 @@ const CreateVoteWorkflow = () => {
     setVoteInfo({ ...voteInfo, activeStep: voteInfo.activeStep - 1 });
   };
 
-  return (
-    <Box
-      sx={{
-        display: "flex",
-        width: "100%",
-        height: "100%",
-      }}
-    >
-      <Box
-        component="nav"
-        sx={{
-          width: { sm: drawerWidth },
-          flexShrink: { sm: 0 },
-        }}
-        aria-label="mailbox folders"
-      >
-        {/* The implementation can be swapped with js to avoid SEO duplication of links. */}
-        <StyledDrawer
-          container={container}
-          variant="temporary"
-          open={mobileOpen}
-          onClose={handleDrawerToggle}
-          ModalProps={{
-            keepMounted: true, // Better open performance on mobile.
-          }}
-          sx={{
-            display: { xs: "block", sm: "none" },
-          }}
-        >
-          {drawer}
-        </StyledDrawer>
-        <StyledDrawer
-          variant="permanent"
-          sx={{
-            display: { xs: "none", sm: "block" },
-          }}
-          open
-        >
-          {drawer}
-        </StyledDrawer>
-      </Box>
-      <Box
-        component="main"
-        sx={{
-          flexGrow: 1,
-          width: { sm: `calc(100% - ${drawerWidth}px)` },
-          height: "100%",
-        }}
-      >
-        <Paper
-          sx={{
-            height: "100%",
-            width: "100%",
-            textAlign: "center",
-          }}
-        >
-          {voteInfo.activeStep === 0 && (
-            <SelectParticipants handleNext={handleNext} />
-          )}
-          {voteInfo.activeStep === 1 && (
-            <SelectCandidates handleNext={handleNext} />
-          )}
-          {voteInfo.activeStep === 2 && (
-            <DatePicker
-              dateValue={dateValue}
-              setDateValue={setDateValue}
-              earliestDate={new Date(new Date().getTime() + DELAY + 60000)} // earliest start date is today
-              endDate={
-                new Date(
-                  new Date().getFullYear() + 7, // set endDate to 7 years in the future
-                  new Date().getMonth(),
-                  new Date().getDate()
-                )
-              }
-              label="Start"
-              handleNext={handleNext}
-              handleBack={handleBack}
-            />
-          )}
-          {voteInfo.activeStep === 3 && (
-            <DatePicker
-              dateValue={dateValue}
-              setDateValue={setDateValue}
-              earliestDate={new Date(voteInfo.startDate.getTime())}
-              endDate={
-                new Date(
-                  voteInfo.startDate.getFullYear() + 7, // set endDate to 7 years in the future
-                  voteInfo.startDate.getMonth(),
-                  voteInfo.startDate.getDate()
-                )
-              }
-              label="End"
-              handleNext={handleNext}
-              handleBack={handleBack}
-            />
-          )}
-          {voteInfo.activeStep === 4 && (
-            <ReviewDetails handleNext={handleNext} handleBack={handleBack} />
-          )}
-          {voteInfo.activeStep === 5 && <Payment handleBack={handleBack} />}
-        </Paper>
-      </Box>
-    </Box>
+  const content = (
+    <div>
+      {voteInfo.activeStep === 0 && (
+        <SelectParticipants handleNext={handleNext} />
+      )}
+      {voteInfo.activeStep === 1 && (
+        <SelectCandidates handleNext={handleNext} />
+      )}
+      {voteInfo.activeStep === 2 && (
+        <DatePicker
+          dateValue={dateValue}
+          setDateValue={setDateValue}
+          earliestDate={new Date(new Date().getTime() + DELAY + 60000)} // earliest start date is today
+          endDate={
+            new Date(
+              new Date().getFullYear() + 7, // set endDate to 7 years in the future
+              new Date().getMonth(),
+              new Date().getDate()
+            )
+          }
+          label="Start"
+          handleNext={handleNext}
+          handleBack={handleBack}
+        />
+      )}
+      {voteInfo.activeStep === 3 && (
+        <DatePicker
+          dateValue={dateValue}
+          setDateValue={setDateValue}
+          earliestDate={new Date(voteInfo.startDate.getTime())}
+          endDate={
+            new Date(
+              voteInfo.startDate.getFullYear() + 7, // set endDate to 7 years in the future
+              voteInfo.startDate.getMonth(),
+              voteInfo.startDate.getDate()
+            )
+          }
+          label="End"
+          handleNext={handleNext}
+          handleBack={handleBack}
+        />
+      )}
+      {voteInfo.activeStep === 4 && (
+        <ReviewDetails handleNext={handleNext} handleBack={handleBack} />
+      )}
+      {voteInfo.activeStep === 5 && <Payment handleBack={handleBack} />}
+    </div>
   );
+  return <WorkflowDrawer steps={steps} content={content} />;
 };
 
 export default CreateVoteWorkflow;
